requirements.txt
Django>=4.0,<5.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
requests>=2.28.0
PyJWT>=2.6.0
